Removed ExternalLinkActionHandler
