+++ lockbox-ios/Presenter/SettingListPresenter.swift
@@ -16,7 +16,6 @@ protocol SettingListViewProtocol: class, AlertControllerView {
 class SettingListPresenter {
     weak private var view: SettingListViewProtocol?
     private let dispatcher: Dispatcher
-    private let linkActionHandler: LinkActionHandler
     private let userDefaultStore: UserDefaultStore
     private let biometryManager: BiometryManager
     private let disposeBag = DisposeBag()
@@ -45,7 +44,7 @@ class SettingListPresenter {
 
     private var setPasscodeButtonObserver: AnyObserver<Void> {
         return Binder(self) { target, _ in
-            target.linkActionHandler.invoke(SettingLinkAction.touchIDPasscode)
+            target.dispatcher.dispatch(action: SettingLinkAction.touchIDPasscode)
             }.asObserver()
     }
 
@@ -83,12 +82,10 @@ class SettingListPresenter {
 
     init(view: SettingListViewProtocol,
          dispatcher: Dispatcher = .shared,
-         linkActionHandler: LinkActionHandler = LinkActionHandler.shared,
          userDefaultStore: UserDefaultStore = .shared,
          biometryManager: BiometryManager = BiometryManager()) {
         self.view = view
         self.dispatcher = dispatcher
-        self.linkActionHandler = linkActionHandler
         self.userDefaultStore = userDefaultStore
         self.biometryManager = biometryManager
     }
